chore: Remove code duplication for api routes
