+++ packages/next/src/server/future/route-modules/helpers/response-handlers.ts
@@ -17,14 +17,6 @@ export function handleBadRequestResponse(): Response {
   return new Response(null, { status: 400 })
 }
 
-export function handleForbiddenResponse(): Response {
-  return new Response(null, { status: 403 })
-}
-
-export function handleNotFoundResponse(): Response {
-  return new Response(null, { status: 404 })
-}
-
 export function handleMethodNotAllowedResponse(): Response {
   return new Response(null, { status: 405 })
 }
